+++ Include/SA/Render/LowLevel/Vulkan/Device/Command/VkCommandBuffer.hpp
@@ -21,7 +21,7 @@ namespace SA::RND::VK
 		void Begin();
 		void End();
 
-		void CopyBuffer(const Buffer& _src, Buffer& _dst, uint32_t _size, uint32_t _srcOffset = 0u, uint32_t _dstOffset = 0u);
+		void CopyBuffer(const Buffer& _src, Buffer& _dst, uint64_t _size, uint64_t _srcOffset = 0u, uint64_t _dstOffset = 0u);
 
 		void Draw(uint32_t _vertexNum, uint32_t _instanceNum, uint32_t _firstVertex, uint32_t _firstInstance);
